监听aria2通知

+++ src/DFApp.Domain/Aria2/Response/TellStatus/UrisItem.cs
@@ -8,14 +8,14 @@ using Volo.Abp.Domain.Entities.Auditing;
 
 namespace DFApp.Aria2.Response.TellStatus
 {
-    public class UrisItem : CreationAuditedAggregateRoot<long>
+    public class UrisItem : CreationAuditedAggregateRoot<short>
     {
         public string Status { get; set; }
 
         public string Uri { get; set; }
 
         public FilesItem FilesItem { get; set; }
-        public long FilesItemId { get; set; }
+        public int FilesItemId { get; set; }
 
     }
 }
